Fix memory leak for storeless measure and nav bug
StorelessMeasureFlowSupervisor used itself as a store reference which
caused a strong reference cycle from the supervisor to the child
supervisor, to the child's store back to the original object. This lead
the stack navigation to call "didDisconnect" on the leaked object on any
navigation movement after the object should already have been dropped.
Removing this cycle lets the StorelessMeasureFlowSupervisor be dropped
when the user cancels the measure edit.

StackNavigation has been updated to auto-drop a stacked delegate if the
delegate object has been dropped and is nil. This lets supervisor
dropping be a valid way to move to the next delegate in the navigiation
stack.